backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["rainline00 <kanjyukaggle@gmail.com>"]
readme = "README.md"
package-mode = false
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.32"
gqlalchemy = "^1.6.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E501", "W391"]
select = ["E", "F", "W", "B", "C"]
line-length = 120
fix = true
src = ["src", "tests"]
